pyproject.toml
[tool.poetry]
name = "bffs"
version = "0.1.0"
description = ""
authors = ["Jeong SeungHo <teapackage16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-cors-headers = "^4.0.0"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"
pre-commit = "^3.3.2"
requests = "^2.31.0"
python-decouple = "^3.8"
django-cron = "^0.6.0"
django-hitcount = "^1.3.5"
djangorestframework = "^3.14.0"
celery = "^5.3.0"
eventlet = "^0.33.3"
django-resized = "^1.0.2"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
